Switch from double to float for drive messages in tests

+++ steering_controllers_library/test/test_steering_controllers_library_ackermann.cpp
@@ -100,7 +100,7 @@ TEST_F(SteeringControllersLibraryTest, test_both_update_methods_for_ref_timeout)
   }
 
   // set command statically
-  const double TEST_LINEAR_VELOCITY_X = 1.5;
+  const float TEST_LINEAR_VELOCITY_X = static_cast<float>(1.5);
   const float TEST_STEERING_ANGLE = static_cast<float>(0.575875);
 
   std::shared_ptr<ControllerAckermannReferenceMsg> msg =
